4.1.6 Разобьёт модуль с утилитами на 3 файла: в одном будут функции для рендеринга, в другом для фильмов, в третеьм функции общего назначения.
