+++ src/utils/render.js
@@ -1,0 +1,38 @@
+export const RenderPosition = {
+  AFTERBEGIN: `afterbegin`,
+  AFTEREND: `afterend`,
+  BEFOREBEGIN: `beforebegin`,
+  BEFOREEND: `beforeend`
+};
+
+export const render = (container, element, place) => {
+  switch (place) {
+    case RenderPosition.AFTERBEGIN: {
+      container.prepend(element);
+      break;
+    }
+    case RenderPosition.AFTEREND: {
+      container.after(element);
+      break;
+    }
+    case RenderPosition.BEFOREBEGIN: {
+      container.before(element);
+      break;
+    }
+    case RenderPosition.BEFOREEND: {
+      container.append(element);
+      break;
+    }
+  }
+};
+
+export const renderTemplate = (container, template, place) => {
+  container.insertAdjacentHTML(place, template);
+};
+
+export const createElement = (template) => {
+  const newElement = document.createElement(`div`);
+  newElement.innerHTML = template;
+
+  return newElement.firstChild;
+};
